move scripts to partials
